[Improvement] Code refractoring | Error handling for users | Sorted array from CoreData
* Now when recruitmentItemsEntityData in pulled from Core Data it is
sorted before presenting.
* Divided Network into APIClient and HttpRequest
* Added ShowsAlert which enables user to see error while fetching data
from server
* Other naming changes and refractoring

+++ iOSRecruitmentTest/Network/RecruitmentItemModel.swift
@@ -8,11 +8,11 @@
 
 import UIKit
 
-struct RecruitmentItemModel: Decodable{
+struct RecruitmentItemModel: Decodable {
     let id: Int
     let name: String
     let description: String
     let icon: String
-    let timestamp: Int
+    let timestamp: Double
     let url: String
 }
